Thesis Deadline Update
Some changes have been made to the project.

All of the base symbols for location are now in and can be input.

Many of the files have more or better documentation.

The regex string for the input has been changed so that the order in which the symbols are input is more intuitive.

The placement of the finger number was supposed to come before the part of the finger referenced; 1ŋr is correct while ŋ1r is not. This has been corrected.